pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "librosa>=0.11.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "reportlab>=4.4.4",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "streamlit>=1.51.0",
    "streamlit-webrtc>=0.63.11",
    "tensorflow>=2.20.0",
]
